cli: Configure App's Reader, Writer, ErrWriter to use OS' stdin, stdout
The underlying library defaults to use stdlib's os.Stdin and os.Stdout
when the fields are not populated. In implementations where OS uses
custom files to represent stdin and stdout, incorrect files are used by
the App.

This commit assigns App's Reader, Writer, ErrWriter to OS's, Stdin,
Stdout. As Risor doesn't have a concept of stderr, stdout is used for
ErrWriter instead.